toast messages

+++ lib/project/admin/services_manage/delete_service.dart
@@ -3,15 +3,13 @@ import 'dart:convert';
 import 'package:conditional_builder_null_safety/conditional_builder_null_safety.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import '../../../Cubit/Admin Level Operation/admin_level_cubit.dart';
-import '../../../main.dart';
-import '/Api/controller/User/work/services_controller.dart';
+import 'package:http/http.dart' as http;
+
 import '/Api/model/name_service.dart';
 import '/project/constant.dart';
 import '/services/choices.dart';
-import 'package:http/http.dart' as http;
-
-import '../../../moh_project/post_moh/login_controller.dart';
+import '../../../Cubit/Admin Level Operation/admin_level_cubit.dart';
+import '../../../main.dart';
 
 class DeleteService extends StatefulWidget {
   const DeleteService({Key? key}) : super(key: key);
@@ -152,22 +150,31 @@ class _DeleteServiceState extends State<DeleteService> {
                                 farmerServices.clear();
                                 bothStreetServices.clear();
 
-
-                                if (snapshot.data!.services.bothStreet.length == 0) {
+                                if (snapshot.data!.services.bothStreet.length ==
+                                    0) {
                                   bothStreetServices.add('no item');
                                 } else {
                                   for (int i = 0;
-                                  i < snapshot.data!.services.bothStreet.length; i++)
-                                  {bothStreetServices.add(snapshot.data!.services.bothStreet[i].name);
+                                      i <
+                                          snapshot
+                                              .data!.services.bothStreet.length;
+                                      i++) {
+                                    bothStreetServices.add(snapshot
+                                        .data!.services.bothStreet[i].name);
                                   }
                                 }
 
-                                if (snapshot.data!.services.woodward.length == 0) {
+                                if (snapshot.data!.services.woodward.length ==
+                                    0) {
                                   woodWardServices.add('no item');
                                 } else {
                                   for (int i = 0;
-                                      i < snapshot.data!.services.woodward.length; i++)
-                                  {woodWardServices.add(snapshot.data!.services.woodward[i].name);
+                                      i <
+                                          snapshot
+                                              .data!.services.woodward.length;
+                                      i++) {
+                                    woodWardServices.add(snapshot
+                                        .data!.services.woodward[i].name);
                                   }
                                 }
                                 //
@@ -236,7 +243,6 @@ class _DeleteServiceState extends State<DeleteService> {
                                             selectedService = servicesList[0];
                                           }
 
-
                                           if (selectedStreet == 'FARMER') {
                                             servicesList =
                                                 List.from(farmerServices);
@@ -250,7 +256,7 @@ class _DeleteServiceState extends State<DeleteService> {
                                           }
                                           // servicesList.clear();
 
-                                          /*
+/*
                             if (selectedStreet=='WoodWard')
                             {
 
@@ -317,9 +323,6 @@ class _DeleteServiceState extends State<DeleteService> {
                               }
                             }
 
- */
-
-/*
                             Navigator.pushReplacement(
                                 context,
                                 MaterialPageRoute(
@@ -402,18 +405,18 @@ class _DeleteServiceState extends State<DeleteService> {
                                               break;
                                             }
                                           }
-                                        }
-                                        else if (selectedStreet == 'BothStreet') {
+                                        } else if (selectedStreet ==
+                                            'BothStreet') {
                                           for (int i = 0;
-                                          i <
-                                              snapshot.data!.services.bothStreet
-                                                  .length;
-                                          i++) {
+                                              i <
+                                                  snapshot.data!.services
+                                                      .bothStreet.length;
+                                              i++) {
                                             if (selectedService ==
                                                 snapshot.data!.services
                                                     .bothStreet[i].name) {
-                                              theId = snapshot
-                                                  .data!.services.bothStreet[i].id;
+                                              theId = snapshot.data!.services
+                                                  .bothStreet[i].id;
 
                                               break;
                                             }
@@ -450,10 +453,10 @@ class _DeleteServiceState extends State<DeleteService> {
                                               //fetchFinal();
                                               // FetchList();
                                               //
-                                              checkServiceDelete(
-                                                  selectedService);
-                                              cubit.deleteService(theId);
-                                              //
+                                              if (checkServiceDelete(
+                                                  selectedService)) {
+                                                cubit.deleteService(theId);
+                                              } //
                                             },
                                             style: ElevatedButton.styleFrom(
                                               minimumSize: const Size(300, 60),
@@ -465,7 +468,7 @@ class _DeleteServiceState extends State<DeleteService> {
                                                             .width *
                                                         0.2,
                                               ),
-                                              primary: const Color.fromARGB(
+                                              backgroundColor: const Color.fromARGB(
                                                   255, 150, 10, 10),
                                               shape: RoundedRectangleBorder(
                                                 borderRadius:
